# pandora_console/api/v2/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Pandora FMS",
    "description": "<p>This is the new API framework for Pandora FMS.</p>\n<p>The old API is deprecated but still functional, and not all old endpoints are supported in the new API, but new endpoints will be added in each release.</p>\n<p>Using this web interface, you can play around and see how it works each endpoint interactively.</p>\n\nMore useful links:\n\n * <a target='_blank' href='https://pandoraopen.io/'>Pandora FMS Licence </a>\n * <a target='_blank' href='https://support.pandoraopen.io'> Pandora FMS Official Support </a>\n * <a target='_blank' href='https://pandoraopen.io/'> Pandora FMS Community </a>\n * <a target='_blank' href='https://pandoraopen.io/'> Vulnerability Disclosure Policy </a>\n * <a target='_blank' href='https://pandoraopen.io/'> Pandora FMS FAQ </a>",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "/api/v2",
      "description": "PandoraFMS API Server"
    }
  ],
  "paths": {
...
    "/event/{idEvent}/comment/{idComment}": {
      "get": {
        "tags": ["Events"],
        "summary": "show field events comments",
        "operationId": "8af1ae411de1815fcab2f36d14097087",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdEvent"
          },
          {
            "$ref": "#/components/parameters/parameterIdEventComment"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseEventComment"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["Events"],
        "summary": "Updates comment for event type",
        "operationId": "37ee3bcc633964ad0cc9449656d31c06",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdEvent"
          },
          {
            "$ref": "#/components/parameters/parameterIdEventComment"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/requestBodyEventComment"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseEventComment"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Events"],
        "summary": "Delete comment for event type object.",
        "operationId": "a8bee70746f69bc9e33430e366c42d96",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdEvent"
          },
          {
            "$ref": "#/components/parameters/parameterIdEventComment"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/successfullyDeleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/event/filter/{idEventFilter}": {
      "get": {
        "tags": ["Events"],
        "summary": "Show eventFilter",
        "operationId": "28583c0a4c279c9098177b00f2ee5d76",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdEventFilter"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseEventFilter"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["Events"],
        "summary": "Updates an eventFilter",
        "operationId": "139dde116c97687e11e446c004575e5c",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdEventFilter"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/requestBodyEventFilter"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseEventFilter"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Events"],
        "summary": "Deletes an eventFilter object.",
        "operationId": "13bbb4cf52878928a5849ed568cf0844",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdEventFilter"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/successfullyDeleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/profile/{idProfile}": {
      "get": {
        "tags": ["Profiles"],
        "summary": "Show profile",
        "operationId": "e66704aa8cb757e23de04ce65c7bf028",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdProfile"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseProfile"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["Profiles"],
        "summary": "Updates an profile",
        "operationId": "723fc236ab3b6069b7dcca9e317c953f",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdProfile"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/requestBodyProfile"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseProfile"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Profiles"],
        "summary": "Deletes an profile object.",
        "operationId": "ca8ec2d296930aac798229f83e93bff8",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdProfile"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/successfullyDeleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/user/{idUser}/login": {
      "get": {
        "tags": ["Users"],
        "summary": "show user when login process",
        "operationId": "1b50ee1984a43d87d69342174571775c",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdUser"
          },
          {
            "$ref": "#/components/parameters/parameterIdUserPass"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseUser"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventComment": {
        "properties": {
          "idEventComment": {
            "description": "id event comment",
            "type": "integer",
            "readOnly": true,
            "nullable": false
          },
          "idEvent": {
            "description": "id event comment",
            "type": "integer",
            "readOnly": true,
            "nullable": false
          },
          "utimestamp": {
            "description": "Event comment utimestamp create",
            "type": "integer",
            "default": null,
            "readOnly": true,
            "example": "1704898868",
            "nullable": true
          },
          "comment": {
            "description": "content of the comment",
            "type": "string",
            "default": null,
            "nullable": false
          },
          "idUser": {
            "description": "User id create comment",
            "type": "string",
            "default": null,
            "readOnly": true,
            "nullable": false
          },
          "action": {
            "description": "content of the action",
            "type": "string",
            "default": null,
            "nullable": false
          }
        },
        "type": "object"
      },
...
      "Profile": {
        "properties": {
          "idProfile": {
            "description": "Id Profile",
            "type": "integer",
            "readOnly": true,
            "nullable": false
          },
          "name": {
            "description": "Name of the profile",
            "type": "string",
            "default": null,
            "nullable": false
          },
          "isAgentView": {
            "description": "Agent Read",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isAgentEdit": {
            "description": "Agent Write",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isAlertEdit": {
            "description": "Alert write",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isUserManagement": {
            "description": "User Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isDbManagement": {
            "description": "Database Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isAlertManagement": {
            "description": "Alert Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isPandoraManagement": {
            "description": "Pandora Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isReportView": {
            "description": "Report view",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isReportEdit": {
            "description": "Report Edit",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isReportManagement": {
            "description": "Report Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isEventView": {
            "description": "Event Read",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isEventEdit": {
            "description": "Event write",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isEventManagement": {
            "description": "Event Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isAgentDisable": {
            "description": "Agent Disable",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isMapView": {
            "description": "Map Read",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isMapEdit": {
            "description": "Map Write",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isMapManagement": {
            "description": "Map Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isVconsoleView": {
            "description": "Visual console Read",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isVconsoleEdit": {
            "description": "Visual console Write",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isVconsoleManagement": {
            "description": "Visual console Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isNetworkConfigView": {
            "description": "Network config read",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isNetworkConfigEdit": {
            "description": "Network config Write",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isNetworkConfigManagement": {
            "description": "Network config Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          }
        },
        "type": "object"
      },
...
    },
    "responses": {
...
      "ResponseEventComment": {
        "description": "Event comment object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EventComment"
            }
          }
        }
      },
...
      "ResponseEventFilter": {
        "description": "EventFilter object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EventFilter"
            }
          }
        }
      },
...
      "ResponseProfile": {
        "description": "Profile object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Profile"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "successfullyDeleted": {
        "description": "Successfully deleted",
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "result": {
                  "type": "string",
                  "default": "Successfully deleted"
                }
              },
              "type": "object"
            }
          }
        }
      },
...
      "ResponseUser": {
        "description": "User object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
...
    },
    "parameters": {
...
      "parameterIdEventComment": {
        "name": "idComment",
        "in": "path",
        "description": "Event comment id",
        "required": true,
        "schema": {
          "type": "integer",
          "default": 1
        }
      },
      "parameterIdEvent": {
        "name": "idEvent",
        "in": "path",
        "description": "Event id",
        "required": true,
        "schema": {
          "type": "integer",
          "default": 1
        }
      },
      "parameterIdEventFilter": {
        "name": "idEventFilter",
        "in": "path",
        "description": "EventFilter id",
        "required": true,
        "schema": {
          "type": "integer",
          "default": 1
        }
      },
...
      "parameterIdProfile": {
        "name": "idProfile",
        "in": "path",
        "description": "Profile id",
        "required": true,
        "schema": {
          "type": "integer",
          "default": 1
        }
      },
...
      "parameterIdUserPass": {
        "name": "password",
        "in": "query",
        "description": "User password",
        "required": true,
        "schema": {
          "type": "string",
          "default": null
        }
      },
      "parameterIdUser": {
        "name": "idUser",
        "in": "path",
        "description": "User id",
        "required": true,
        "schema": {
          "type": "string",
          "default": "admin"
        }
      },
...
    },
    "requestBodies": {
...
      "requestBodyEventComment": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EventComment"
            }
          }
        }
      },
...
      "requestBodyEventFilter": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/EventFilter"
            }
          }
        }
      },
...
      "requestBodyProfile": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Profile"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "in": "header",
        "bearerFormat": "pandoraBearer",
        "scheme": "bearer"
      }
    }
  },
...
}